requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
python-telegram-bot==21.3
lxml==5.2.1
python-dotenv==1.0.1
playwright
